chore: resolved comments
