+++ messages/assess.json
@@ -13,6 +13,7 @@
   "commandDescription": "If your Omnistudio components are on custom data model, assess the migration of the components from Omnistudio managed package runtime to Omnistudio standard runtime. If your Omnistudio components are on standard data model, assess the updates made to the component names and references.See: https://help.salesforce.com/s/articleView?id=xcloud.os_omnistudio_naming_conventions.htm&type=5",
   "errorNoOrgResults": "No records found in the org '%s' to migrate.",
   "onlyFlagDescription": "Migrate a single component: os | ip | fc | dm | autonumber | cl",
+  "invalidOnlyFlag": "Invalid parameter entered. Select a valid parameter from these options: os | ip | fc | dm | autonumber | cl",
   "couldNotTruncate": "We couldn't clear all %s records.",
   "couldNotTruncateOmnniProcess": "We couldn't clear all records of your %s because the %s is referenced in an Omniscript or a Flexcard.",
   "invalidOrRepeatingOmniscriptElementNames": "Omniscript with invalid or duplicate element names found. Rename your Omniscript elements and try again.",
