+++ src/commands/omnistudio/migration/migrate.ts
@@ -75,6 +75,14 @@ export default class Migrate extends OmniStudioBaseCommand {
     }),
   };
 
+  // OmniStudio components that don't need migration logging when metadata API is enabled
+  private readonly OMNISTUDIO_COMPONENTS_FOR_LOGGING = [
+    Constants.Omniscript,
+    Constants.IntegrationProcedure,
+    Constants.Flexcard,
+    Constants.DataMapper,
+  ];
+
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   public async run(): Promise<any> {
     Logger.initialiseLogger(this.ux, this.logger, 'migrate', this.flags.verbose);
@@ -389,22 +397,10 @@ export default class Migrate extends OmniStudioBaseCommand {
     for (const cls of migrationObjects) {
       try {
         const componentName = cls.getName();
-        const shouldSkipLogging = this.shouldSkipMigrationLogging(componentName);
 
-        if (!shouldSkipLogging) {
-          Logger.log(messages.getMessage('migratingComponent', [componentName]));
-        }
-        debugTimer.lap('Migrating: ' + componentName);
+        this.logMigrationStart(componentName, debugTimer);
         const results = await cls.migrate();
-
-        // Skip success/failure logs for components that don't need migration when metadata API is enabled
-        if (!shouldSkipLogging) {
-          if (results.some((result) => result?.errors?.length > 0) && !isStandardDataModelWithMetadataAPIEnabled()) {
-            Logger.error(messages.getMessage('migrationFailed', [componentName]));
-          } else {
-            Logger.log(messages.getMessage('migrationCompleted', [componentName]));
-          }
-        }
+        this.logMigrationComplete(componentName, results);
 
         objectMigrationResults = objectMigrationResults.concat(
           results.map((r) => {
@@ -435,6 +431,33 @@ export default class Migrate extends OmniStudioBaseCommand {
   }
 
   /**
+   * Log the start of component migration and track timing
+   */
+  private logMigrationStart(componentName: string, debugTimer: DebugTimer): void {
+    const shouldSkipLogging = this.shouldSkipMigrationLogging(componentName);
+
+    if (!shouldSkipLogging) {
+      Logger.log(messages.getMessage('migratingComponent', [componentName]));
+      debugTimer.lap('Migrating: ' + componentName);
+    }
+  }
+
+  /**
+   * Log the completion or failure of component migration
+   */
+  private logMigrationComplete(componentName: string, results: MigrationResult[]): void {
+    const shouldSkipLogging = this.shouldSkipMigrationLogging(componentName);
+
+    if (!shouldSkipLogging) {
+      if (results.some((result) => result?.errors?.length > 0) && !isStandardDataModelWithMetadataAPIEnabled()) {
+        Logger.error(messages.getMessage('migrationFailed', [componentName]));
+      } else {
+        Logger.log(messages.getMessage('migrationCompleted', [componentName]));
+      }
+    }
+  }
+
+  /**
    * Check if migration logging should be skipped for a component
    * Returns true if Metadata API is enabled and the component doesn't need migration
    */
@@ -444,14 +467,9 @@ export default class Migrate extends OmniStudioBaseCommand {
     }
 
     // Skip logging for OmniStudio components that don't need migration when metadata API is enabled
-    const omnistudioComponents = [
-      Constants.Omniscript,
-      Constants.IntegrationProcedure,
-      Constants.Flexcard,
-      Constants.DataMapper,
-    ];
-
-    return omnistudioComponents.some((component) => componentName.toLowerCase().includes(component.toLowerCase()));
+    return this.OMNISTUDIO_COMPONENTS_FOR_LOGGING.some((component) =>
+      componentName.toLowerCase().includes(component.toLowerCase())
+    );
   }
 
   /**
